+++ src/module/rolls/index.ts
@@ -1,3 +1,3 @@
-export { IronswornRoll } from './roll'
+export { IronswornRoll } from './ironsworn-roll'
 export { IronswornPrerollDialog } from './preroll-dialog'
 export { IronswornRollChatMessage } from './chat-message'
